requirements_ai.txt
# ===== AI / ML TRAINING DEPENDENCIES =====
# Core ML Libraries
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
xgboost==3.2.0
imbalanced-learn==0.14.1
scipy==1.17.1

# Visualization
matplotlib==3.10.8
seaborn==0.13.2
pillow==12.1.1

# Deep Learning (Optional but included)
transformers==5.2.0
torch==2.10.0
tf-keras==2.15.0

# Configuration
python-dotenv==1.2.2
PyYAML==6.0.3

# Utilities
tqdm==4.67.3
joblib==1.5.3

# Auto-installed Dependencies
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
typing_extensions==4.15.0
tzdata==2025.3
setuptools==82.0.0
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.5.0
filelock==3.24.3
packaging==26.0
python-dateutil==2.9.0.post0
six==1.17.0
threadpoolctl==3.6.0
click==8.3.1
colorama==0.4.6
decorator==5.2.1
pooch==1.9.0
lazy_loader==0.4
msgpack==1.1.2
networkx==3.6.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pyparsing==3.3.2
Jinja2==3.1.6
MarkupSafe==3.0.3
soupsieve==2.8.3
cffi==2.0.0
pycparser==3.0
llvmlite==0.46.0
numba==0.64.0
RapidFuzz==3.14.3
regex==2026.2.28
httpx==0.28.1
h11==0.16.0
httpcore==1.0.9
anyio==4.12.1
future==1.0.0
sympy==1.14.0
mpmath==1.3.0
Pygments==2.19.2
rich==14.3.3
markdown-it-py==4.0.0
mdurl==0.1.2
fsspec==2026.2.0
platformdirs==4.9.2
